change(env): Improved environment parsing
getEnvRaw: `getEnvRaw` is essentially the same as `getEnv`, but much more safe as it doesn't automatically convert types - parsing is left to the consuming code.
parseBool: `parseBool` is a much more accepting function for parsing booleans. `true`, `1`, `yes`, and `on` are treated as `true`, and conversely, `false`, `0`, `no`, and `off` are treated as `false`.