change: Reworded footer in shared documents, updated link
rewording: Previously, the footer said "Built with", which makes it seem too technical. This just changes it to "Written with", which is a more accurate portrayal. Fixes #154.
updated-link: The new link is https://github.com/notea-org/notea, which is preferred to the previous link (https://cinwell.com/notea) if a website for Notea is necessary.